+++ plugin/src/adminPanel/tableList.php
@@ -24,11 +24,11 @@ function show_tableList()
     wp_enqueue_script("sort_script", plugins_url("script/sort.js", __FILE__));
 
 
-    if (isset($_POST["tableToDelete"])) {
+    if (isset($_POST["tableToDelete"]) && current_user_can("tv_deleteTables")) {
         deleteTable($_POST["tableToDelete"]);
     }
 
-    if (isset($_POST["tableToEdit"]) && isset($_POST["title"]) && isset($_POST["numberOfSeats"])) {
+    if (current_user_can("tv_editTables") && isset($_POST["tableToEdit"]) && isset($_POST["title"]) && isset($_POST["numberOfSeats"])) {
         if (empty($_POST["tableToEdit"]) || empty($_POST["title"]) || empty($_POST["numberOfSeats"])) {
             echo '<p class="formError">Bitte fülle alle Pflichtfelder aus!</p>';
         } else {
@@ -50,7 +50,7 @@ function show_tableList()
 
 <div id="main">
     <h1 class="inline">Alle Tische</h1>
-    <a href="admin.php?page=addtable" class="btn">Neuen Tisch erstellen</a>
+    <?php if(current_user_can("tv_addTables")) echo '<a href="admin.php?page=addtable" class="btn">Neuen Tisch erstellen</a>' ?>
     <form method="post">
         <table class="content">
             <tr id="head">
@@ -85,11 +85,18 @@ function show_tableList()
         echo '<td class="m_numberOfSeats">'.$r["seats"].'</td>';
         echo '<td class="m_isDisabled">'.($r["isDisabled"] ? "ja" : "nein").'</td>';
 
-        echo '<td><button type="submit" name="tableToDelete" class="edit" id="deleteBtn_'.$r["id"].'" value="'.$r["id"].'" onclick="return confirm(\'Willst du diesen Eintrag wirklich löschen?\');"><i class="fa fa-trash"></i></button>';
+        echo '<td>';
+        if(current_user_can("tv_deleteTables")) {
+            echo '<button type="submit" name="tableToDelete" class="edit" id="deleteBtn_'.$r["id"].'" value="'.$r["id"].'" onclick="return confirm(\'Willst du diesen Eintrag wirklich löschen?\');"><i class="fa fa-trash"></i></button>';
+        }
 
-        echo ' <button type="button" id="editBtn_'.$r["id"].'" class="edit" onclick="edit('.$r["id"].')"><i class="fa fa-pencil"></i></button>';
-        echo ' <button type="submit"  id="saveBtn_'.$r["id"].'" class="hidden edit" name="tableToEdit" value="'.$r["id"].'"><i class="fa fa-floppy-o"></i></button>';
-        echo ' <button type="button"  id="cancelBtn_'.$r["id"].'" onclick="cancelEdit()" class="hidden edit"><i class="fa fa-times"></i></button></td>';
+        if(current_user_can("tv_editTables")) {
+            echo ' <button type="button" id="editBtn_'.$r["id"].'" class="edit" onclick="edit('.$r["id"].')"><i class="fa fa-pencil"></i></button>';
+            echo ' <button type="submit"  id="saveBtn_'.$r["id"].'" class="hidden edit" name="tableToEdit" value="'.$r["id"].'"><i class="fa fa-floppy-o"></i></button>';
+            echo ' <button type="button"  id="cancelBtn_'.$r["id"].'" onclick="cancelEdit()" class="hidden edit"><i class="fa fa-times"></i></button>';
+        }
+        echo '</td>';
+        
 
         echo '</tr>';
     } ?>
